Fix Bugs
- The FSM now play every state. Before, it never enters in "attack" state.
- Br1e & CyberRoach fixed sizes

+++ Assets/Scripts/Entities/Enemies/EnemyBasicParameters.cs
@@ -11,7 +11,7 @@ namespace TOM.Enemy
         public int powerAttack;
         public float powerAttackCoolDown;
         public int powerAttackChance;
-        public int attackRadius;
+        public float attackRadius;
         public float stunTimeInms;
         public int movementSpeed;
         public float hurtTime;
